# apps/api/src/lib/dto.ts
/**
 * Tenant-facing response mappers.
 *
 * Every tenant API response goes through one of these before c.json() so we
 * never leak underlying payment-processor identifiers (currently Wingspan).
 *
 * Rules:
 *   - Internal DB columns named `wingspan_*` MUST NOT appear in tenant output.
 *   - The contractor↔entity engagement (formerly exposed as
 *     `wingspanPayerPayeeEngagementId`) is surfaced as `engagementId` —
 *     SlyncPay's own UUID, which is what tenants reference on payables.
 *   - DB column names stay as-is internally; admin endpoints still see them.
 */

// ─── Types (loose) ────────────────────────────────────────────────────────────

type AnyRow = Record<string, unknown>;

export interface WorkerDTO {
  id: string;
  externalId: string;
  email: string;
  firstName: string | null;
  lastName: string | null;
  onboardingStatus: string;
  metadata?: unknown;
  w9SeededData?: unknown;
  ssnLast4?: string | null;
  createdAt: Date | string;
  updatedAt?: Date | string;
}

export interface EngagementDTO {
  id: string;
  workerId: string;
  entityId: string;
  engagementId: string; // SlyncPay's UUID — used on payables
  status: string;
  type: "contractor" | "employee";
  entityName?: string | null;
  createdAt: Date | string;
}

export interface EntityDTO {
  id: string;
  name: string;
  einLast4: string | null;
  state: string | null;
  status: string;
  taxType: "w2" | "1099";
  createdAt: Date | string;
  updatedAt?: Date | string;
}

export interface PayableDTO {
  id: string;
  entityId: string;
  workerId: string;
  engagementId?: string;
  externalReferenceId: string | null;
  amountCents: number;
  status: string;
  lineItems?: unknown;
  dueDate: string;
  disbursementId?: string | null;
  createdAt: Date | string;
  updatedAt?: Date | string;
  paidAt?: Date | string | null;
}

export interface DisbursementDTO {
  id: string;
  entityId: string;
  status: string;
  totalPayablesCount: number;
  totalAmountCents: number;
  initiatedAt: Date | string;
  completedAt?: Date | string | null;
  failureReason?: string | null;
  submittedToProcessor?: boolean;
}

// ─── Mappers ──────────────────────────────────────────────────────────────────

export function toWorkerDTO(r: AnyRow): WorkerDTO {
  return {
    id: r["id"] as string,
    externalId: r["externalId"] as string,
    email: r["email"] as string,
    firstName: (r["firstName"] as string | null) ?? null,
    lastName: (r["lastName"] as string | null) ?? null,
    onboardingStatus: r["onboardingStatus"] as string,
    ...(r["metadata"] !== undefined ? { metadata: r["metadata"] } : {}),
    ...(r["w9SeededData"] !== undefined ? { w9SeededData: r["w9SeededData"] } : {}),
    ...(r["ssnLast4"] !== undefined ? { ssnLast4: (r["ssnLast4"] as string | null) ?? null } : {}),
    createdAt: r["createdAt"] as Date | string,
    ...(r["updatedAt"] !== undefined ? { updatedAt: r["updatedAt"] as Date | string } : {}),
  };
}

export function toEngagementDTO(
  r: AnyRow,
  opts: { entityName?: string | null } = {},
): EngagementDTO {
  return {
    id: r["id"] as string,
    workerId: r["workerId"] as string,
    entityId: r["entityId"] as string,
    engagementId: r["id"] as string, // intentional: tenant-facing ID === SlyncPay UUID
    status: r["status"] as string,
    type: ((r["type"] as string) ?? "contractor") as "contractor" | "employee",
    ...(opts.entityName !== undefined ? { entityName: opts.entityName } : {}),
    createdAt: r["createdAt"] as Date | string,
  };
}

export function toEntityDTO(r: AnyRow): EntityDTO {
  return {
    id: r["id"] as string,
    name: r["name"] as string,
    einLast4: (r["einLast4"] as string | null) ?? null,
    state: (r["state"] as string | null) ?? null,
    status: r["status"] as string,
    taxType: ((r["taxType"] as string) ?? "1099") as "w2" | "1099",
    createdAt: r["createdAt"] as Date | string,
    ...(r["updatedAt"] !== undefined ? { updatedAt: r["updatedAt"] as Date | string } : {}),
  };
}

export function toPayableDTO(r: AnyRow): PayableDTO {
  return {
    id: r["id"] as string,
    entityId: r["entityId"] as string,
    workerId: r["workerId"] as string,
    ...(r["engagementId"] !== undefined ? { engagementId: r["engagementId"] as string } : {}),
    externalReferenceId: (r["externalReferenceId"] as string | null) ?? null,
    amountCents: r["amountCents"] as number,
    status: r["status"] as string,
    ...(r["lineItems"] !== undefined ? { lineItems: r["lineItems"] } : {}),
    dueDate: r["dueDate"] as string,
    ...(r["disbursementId"] !== undefined ? { disbursementId: r["disbursementId"] as string | null } : {}),
    createdAt: r["createdAt"] as Date | string,
    ...(r["updatedAt"] !== undefined ? { updatedAt: r["updatedAt"] as Date | string } : {}),
    ...(r["paidAt"] !== undefined ? { paidAt: r["paidAt"] as Date | string | null } : {}),
  };
}

export function toDisbursementDTO(r: AnyRow): DisbursementDTO {
  const batchId = r["wingspanBulkBatchId"];
  return {
    id: r["id"] as string,
    entityId: r["entityId"] as string,
    status: r["status"] as string,
    totalPayablesCount: r["totalPayablesCount"] as number,
    totalAmountCents: Number(r["totalAmountCents"] ?? 0),
    initiatedAt: r["initiatedAt"] as Date | string,
    ...(r["completedAt"] !== undefined ? { completedAt: r["completedAt"] as Date | string | null } : {}),
    ...(r["failureReason"] !== undefined ? { failureReason: (r["failureReason"] as string | null) ?? null } : {}),
    submittedToProcessor: typeof batchId === "string" && batchId.length > 0,
  };
}