# apps/web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlyncPay API",
    "version": "2.0.0",
...
  },
  "servers": [
    { "url": "https://slyncpay-api.onrender.com", "description": "Production" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "validation_error" },
          "message": { "type": "string", "example": "Field xyz is required" },
          "statusCode": { "type": "integer", "example": 400 }
        }
      },
      "W9Prefill": {
        "type": "object",
        "description": "Fields used to pre-fill the worker's Wingspan W-9 onboarding form. Address fields and country pre-fill the W-9 step; the rest (middleName, jobTitle, dateOfBirth, phone) are stored on the worker record but the worker still enters them in the form themselves.",
        "properties": {
          "middleName": { "type": "string", "maxLength": 100 },
          "jobTitle": { "type": "string", "maxLength": 200, "description": "e.g. 'Travel nurse'." },
          "dateOfBirth": { "type": "string", "format": "date", "description": "YYYY-MM-DD." },
          "phone": { "type": "string", "maxLength": 30 },
          "addressLine1": { "type": "string", "maxLength": 200 },
          "addressLine2": { "type": "string", "maxLength": 200 },
          "city": { "type": "string", "maxLength": 100 },
          "state": { "type": "string", "minLength": 2, "maxLength": 2, "description": "Two-letter US state code." },
          "postalCode": { "type": "string", "maxLength": 20 },
          "country": { "type": "string", "minLength": 2, "maxLength": 2, "default": "US" }
        }
      },
      "Worker": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "externalId": { "type": "string", "description": "Your internal identifier for this worker." },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "onboardingStatus": {
            "type": "string",
            "enum": ["invited", "w9_pending", "payout_pending", "active", "inactive"]
          },
          "w9SeededData": { "anyOf": [{ "$ref": "#/components/schemas/W9Prefill" }, { "type": "null" }] },
          "ssnLast4": { "type": ["string", "null"], "description": "Last 4 digits of stored SSN, if any." },
          "metadata": { "type": "object", "additionalProperties": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Entity": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "einLast4": { "type": ["string", "null"] },
          "state": { "type": ["string", "null"] },
          "status": { "type": "string", "enum": ["pending", "active", "suspended"] },
          "taxType": { "type": "string", "enum": ["1099", "w2"], "description": "Determines whether this entity runs 1099 contractor payments or W-2 employee payroll. Locked at create time." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "LineItem": {
        "type": "object",
        "required": ["amountCents"],
        "properties": {
          "description": { "type": "string", "maxLength": 500 },
          "amountCents": { "type": "integer", "minimum": 1 },
          "quantity": { "type": "number", "minimum": 0 },
          "unit": { "type": "string", "maxLength": 50 }
        }
      },
      "Payable": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "entityId": { "type": "string", "format": "uuid" },
          "workerId": { "type": "string", "format": "uuid" },
          "engagementId": { "type": "string", "format": "uuid" },
          "externalReferenceId": { "type": ["string", "null"], "description": "Your reference (e.g. invoice or shift ID)." },
          "amountCents": { "type": "integer", "minimum": 1 },
          "status": { "type": "string", "enum": ["draft", "pending", "processing", "paid", "failed", "cancelled"] },
          "dueDate": { "type": "string", "format": "date" },
          "lineItems": { "type": "array", "items": { "$ref": "#/components/schemas/LineItem" } },
          "disbursementId": { "type": ["string", "null"], "format": "uuid", "description": "Set once the payable has been swept into a disbursement batch." },
          "paidAt": { "type": ["string", "null"], "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Disbursement": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "entityId": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["processing", "completed", "failed", "partial"] },
          "totalPayablesCount": { "type": "integer" },
          "totalAmountCents": { "type": "integer" },
          "initiatedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": ["string", "null"], "format": "date-time" },
          "failureReason": { "type": ["string", "null"] },
          "submittedToProcessor": { "type": "boolean", "description": "True once the batch has been accepted by the payment processor." }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "hasMore": { "type": "boolean" }
        }
      },
...
    }
  },
...
  "paths": {
    "/v1/workers": {
      "post": {
        "tags": ["Workers"],
        "summary": "Create a worker",
        "description": "Onboards a new worker under your tenant. Pass any data you already have (name, address, SSN) and we'll seed the Wingspan W-9 onboarding form so the worker confirms instead of re-entering it. The worker receives an embedded onboarding URL — drop it in an iframe or email it to them.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["externalId", "email"],
                "properties": {
                  "externalId": { "type": "string", "maxLength": 100, "description": "Your stable identifier (e.g. 'nurse-001'). Must be unique within your tenant + environment." },
                  "email": { "type": "string", "format": "email" },
                  "firstName": { "type": "string", "maxLength": 100 },
                  "lastName": { "type": "string", "maxLength": 100 },
                  "metadata": { "type": "object", "additionalProperties": true },
                  "w9Prefill": { "$ref": "#/components/schemas/W9Prefill" },
                  "ssn": { "type": "string", "description": "9-digit SSN/ITIN (with or without dashes). Encrypted at rest, never returned. Only `ssnLast4` is exposed on the worker DTO." }
                }
              },
              "example": {
                "externalId": "nurse-001",
                "email": "jane@example.com",
                "firstName": "Jane",
                "lastName": "Doe",
                "ssn": "123-45-6789",
                "w9Prefill": {
                  "dateOfBirth": "1985-04-12",
                  "phone": "+15125551234",
                  "jobTitle": "Travel nurse",
                  "addressLine1": "100 Main St",
                  "city": "Austin",
                  "state": "TX",
                  "postalCode": "78701",
                  "country": "US"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Worker" },
                    {
                      "type": "object",
                      "properties": {
                        "embeddedOnboardingUrl": {
                          "type": ["string", "null"],
                          "description": "Iframe-safe URL that walks the worker through W-9, payout setup, and identity. Expires 60 min after issuance."
                        },
                        "embeddedOnboardingExpiresAt": { "type": ["string", "null"], "format": "date-time" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": { "description": "External ID already used", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Tenant provisioning is not complete yet (or sandbox not configured)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["Workers"],
        "summary": "List workers",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["invited", "w9_pending", "payout_pending", "active", "inactive"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Worker" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/entities": {
      "get": {
        "tags": ["Entities"],
        "summary": "List your entities",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Entity" } } } }
          }
        }
      },
      "post": {
        "tags": ["Entities"],
        "summary": "Create an entity",
        "description": "Creates a new legal entity (one per EIN). Provisioning runs asynchronously — poll `GET /v1/entities/{id}/provisioning-status` until the entity's `status` flips to `active` before paying workers against it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "ein"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 },
                  "ein": { "type": "string", "pattern": "^\\d{2}-\\d{7}$", "description": "Format XX-XXXXXXX." },
                  "state": { "type": "string", "minLength": 2, "maxLength": 2 },
                  "taxType": { "type": "string", "enum": ["1099", "w2"], "default": "1099", "description": "Locks the entity's classification. W-2 entities accept worker engagements but payable creation returns 501 until the W-2 runtime ships." }
                }
              },
              "example": { "name": "NurseIO, Inc.", "ein": "83-3439766", "state": "AZ", "taxType": "1099" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created (provisioning runs in background)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Entity" } } } },
          "402": { "description": "Plan entity limit reached", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/v1/payables": {
      "post": {
        "tags": ["Payables"],
        "summary": "Create a payable (queued for next disbursement)",
        "description": "Creates a pending payable for the given worker + entity. It will be paid the next time you trigger a disbursement for that entity.\n\nContractor must have `onboardingStatus = active`; otherwise returns 400 (`validation_error`).\n\n**Idempotency-Key header required.**",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["entityId", "workerId", "amountCents", "dueDate", "lineItems"],
                "properties": {
                  "entityId": { "type": "string", "format": "uuid" },
                  "workerId": { "type": "string", "format": "uuid" },
                  "externalReferenceId": { "type": "string", "maxLength": 200 },
                  "amountCents": { "type": "integer", "minimum": 1 },
                  "dueDate": { "type": "string", "format": "date" },
                  "lineItems": { "type": "array", "items": { "$ref": "#/components/schemas/LineItem" }, "minItems": 1 },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              },
              "example": {
                "entityId": "9e1d…-…",
                "workerId": "7c3f…-…",
                "externalReferenceId": "SHIFT-9021",
                "amountCents": 45000,
                "dueDate": "2026-05-15",
                "lineItems": [{ "description": "Shift on 2026-05-12", "amountCents": 45000 }]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Payable" } } } },
          "400": { "description": "Worker not active, no engagement, or other validation failure", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["Payables"],
        "summary": "List payables",
        "parameters": [
          { "name": "entityId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "workerId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "pending", "processing", "paid", "failed", "cancelled"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Payable" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/disbursements": {
      "post": {
        "tags": ["Disbursements"],
        "summary": "Trigger a batch disbursement",
        "description": "Pays out all pending payables for the given entity in a single batch.\n\n**Idempotency-Key header required.**",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "required": ["entityId"], "properties": { "entityId": { "type": "string", "format": "uuid" } } },
              "example": { "entityId": "9e1d…-…" }
            }
          }
        },
        "responses": {
          "201": { "description": "Disbursement initiated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Disbursement" } } } },
          "400": { "description": "No pending payables for this entity", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "get": {
        "tags": ["Disbursements"],
        "summary": "List disbursements",
        "parameters": [
          { "name": "entityId", "in": "query", "schema": { "type": "string", "format": "uuid" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["processing", "completed", "failed", "partial"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Disbursement" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}